+++ src/game/server/entities/pickup_drop.h
@@ -14,7 +14,7 @@
 class CPickupDrop : public CEntity
 {
 public:
-	CPickupDrop(CGameWorld *pGameWorld, int Type, int Owner, int Direction, int Weapon = WEAPON_GUN, int Lifetime = 300, int Bullets = -1, bool SpreadWeapon = false, bool Jetpack = false);
+	CPickupDrop(CGameWorld *pGameWorld, int Type, int Owner, float Direction, int Weapon = WEAPON_GUN, int Lifetime = 300, int Bullets = -1, bool SpreadWeapon = false, bool Jetpack = false);
 
 	virtual void Reset(bool EreaseWeapon = true, bool Picked = false);
 	virtual void Tick();
